+++ src/main/java/com/transport/management/modules/usuario/UsuarioController.java
@@ -11,8 +11,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.transport.management.services.UsuarioService;
-
 @RestController
 @RequestMapping("/usuarios")
 public class UsuarioController {
